Setup
Adding a setup installation file for Windows computers. This installs the needed dependencies and links to make a working version of project in Angular CLI tool. It also performs an audit and patches any vulnerabilities. Make sure to run each semester. This assumes Node.js is already installed with the correct system variable paths.